+++ provider/tokyo.js
@@ -1,5 +1,0 @@
-const url = process.env.SURGIO_SUBSCRIPTION_URL;
-
-module.exports = url
-  ? { type: 'shadowsocks_subscription', url }
-  : { type: 'custom', nodeList: [] };